requirements.txt
Django==3.2.3
selenium==3.141.0
gunicorn==20.1.0
Fabric3==1.14.post1